backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-Session==0.5.0
scikit-learn==1.4.2
numpy==1.26.4
pandas==2.2.1
gunicorn==21.2.0
Werkzeug==2.3.7
python-dotenv==1.0.0
gunicorn==21.2.0

Flask
Flask-CORS
numpy
pandas
scikit-learn
tensorflow
pymongo
bcrypt
